refactoring CDUD po supplier dan do catering

+++ app/Http/Controllers/API/InternalOrder/DOCateringController.php
@@ -24,15 +24,25 @@ class DOCateringController extends Controller
             'limit' => ['nullable', 'integer'],
             'search' => ['nullable', 'string'],
             'paginate' => ['nullable', 'in:0,1'],
+            'status' => ['nullable', 'in:draft,submit'],
         ]);
         $search = $request->search;
         $paginate = $request->input('paginate', 1);
         $limit = $request->input('limit', 10);
+        $status = $request->status;
 
         $do_catering = DOCatering::when($search, function ($query, string $search) {
-                                            $query->where('do_number', 'like', '%'.$search.'%');
-                                        })
-                                        ->orderBy('created_at', 'DESC');
+                                    $query->where(function ($query2) use ($search) {
+                                        $query2->where('do_number', 'like', '%'.$search.'%')
+                                        ->orWhereHas('po_supplier_catering', function ($sub_query) use ($search) {
+                                            $sub_query->where('po_number', 'like', '%'.$search.'%');
+                                        });
+                                    });
+                                })
+                                ->when($status, function ($query, string $status) {
+                                    $query->where('status', $status);
+                                })
+                                ->orderBy('created_at', 'DESC');
 
         $result = $paginate ? $do_catering->paginate($limit) : $do_catering->get();
 
@@ -40,39 +50,7 @@ class DOCateringController extends Controller
             DOCateringResource::collection($result)->response()->getData(true),
             'success get do catering data'
         );
-    }
-
-    public function store(DOCateringRequest $request) 
-    {
-        $input = $request->except([
-            'item_product'
-        ]);
-        
-        $last_number = $this->last_number();
-        $input['created_by'] = Auth::user()->id;
-        $input['serial_number'] = $last_number;
-        $input['do_number'] = $last_number .'/SBL/DOC/' . DateHelpers::monthToRoman(Carbon::now()->month) .'/'. Carbon::now()->year;
-
-        // database transaction for do catering and item data
-        $result = DB::transaction(function () use ($input, $request) {
-            // store do catering data
-            $do_catering = DOCatering::create($input);
-
-            // store item product data
-            foreach($request->item_product as $item_product) {
-                $item_product['reference_type'] = 'App\Models\DOCatering';
-                $item_product['reference_id'] = $do_catering->id;
-                SelectItemProduct::create($item_product);
-            }
-
-            return $do_catering;
-        });
-
-        return ResponseFormatter::success(
-            new DOCateringDetailResource($result),
-            'success create do catering data'
-        );
-    }    
+    }  
 
     public function show(DOCatering $do_catering)
     {
@@ -84,15 +62,15 @@ class DOCateringController extends Controller
 
     public function update(DOCateringRequest $request, DOCatering $do_catering)
     {
-        $input = $request->except([
-            'item_product'
-        ]);
+        if($do_catering->status == 'submit')
+        {
+            return ResponseFormatter::errorValidation([
+                'do_catering_id' => ['cannot update this data because the status has already been submitted']
+            ], 'update do cateirng data failed', 422);
+        }
 
         // database transaction for do catering and item data
-        $result = DB::transaction(function () use ($input, $request, $do_catering) {
-            // store do catering data
-            $do_catering->update($input);
-
+        $result = DB::transaction(function () use ($request, $do_catering) {
             // delete do catering item product
             $do_catering->item_product()->delete();
 
@@ -112,32 +90,19 @@ class DOCateringController extends Controller
         );
     }
 
-    public function destroy(DOCatering $do_catering)
+    public function update_status(Request $request, DOCatering $do_catering)
     {
-        DB::transaction(function () use ($do_catering) {
-            // delete attachment file
-            $files = $do_catering->attachment_file()->pluck('file')->toArray();
-            Storage::disk('local')->delete($files);    
-            $do_catering->attachment_file()->delete();
+        $request->validate([
+            'status' => ['required', 'in:draft,submit'],
+        ]);
 
-            // delete item product
-            $do_catering->item_product()->delete();
-
-            // delete catering po
-            $do_catering->delete();
-        });
-
+        $input = $request->only('status');
+        $do_catering->update($input);
+        
         return ResponseFormatter::success(
-            null,
-            'success delete do catering data'
+            new DOCateringDetailResource($do_catering),
+            'success update status do catering data'
         );
     }
 
-    public function last_number()
-    {
-         $last_number = DOCatering::whereYear('created_at', Carbon::now()->year)
-                                ->whereMonth('created_at', Carbon::now()->month)
-                                ->max('serial_number');
-        return $last_number + 1;
-    }
 }
